Faster PPM repo update.
Assume that packages from a repo don't change if their name and version number stay the same.
Running 'ppm repo sync -force' will continue to update as before (always clearing out
all packages during the update).

The update process became about twice as fast when I measured it on my Mac (with 2600 packages
in the repo).  Without the new index we ended up twice as slow if we tried to check if the
packages was already there.

[git-p4: depot-paths = "//depot/main/Camel/src/as-mod/ActivePerl-PPM/main/": change = 286674]