Tweak the LWP heuristics for when to consider a response expired.
If the server does not send a Cache-Control or Expires header to
override, this change will make sure that we never check a repo more
than once every day.  If the Last-Modification date indicate that
the packlist has not changed in a long time we might only verify that
it's still up-to-date once every week.

To have effect this change requires LWP-5.820, which I will set up
as soon as it has propagated to our CPAN mirror.

Fixes http://bugs.activestate.com/show_bug.cgi?id=80313

[git-p4: depot-paths = "//depot/main/Camel/src/as-mod/ActivePerl-PPM/main/": change = 287937]