Abort installs when it contains packages marked as uninstallable.
Allow --force to bypass this test so that ppm still tries the install.
This is never expected to actually succeed in installing, but at least
it allow us to test that the server side also implements similar
restrictions and will provide an escape if we ever decided to implement
selective restrictions.

[git-p4: depot-paths = "//depot/main/Camel/src/as-mod/ActivePerl-PPM/main/": change = 291665]