Add sqlite index on the feature(name,role) and feature(package_id)
This makes it much faster to determine if a feature is provided by an
area, which is called quite a lot when trying to resolve what packages
to install.  The feature(package_id) makes it much faster to load
packages from the repo database.

For my Catalyst install test case (installing Catalyst-Devel,
Catalyst-Runtime and the other modules required by the Stackato
catalyst tutorials -- 123 packages installed) this dropped the
time required to evaluate package_missing() from 27s to 1.1s.
That's 25x faster.

Downside is that the initial 'repo sync' now takes longer.
Time spent updating the database went from 17s to 20s.

[git-p4: depot-paths = "//depot/main/Camel/src/as-mod/ActivePerl-PPM/main/": change = 295445]