Drop feature indexes before the 'repo sync' updates the database.
On OS/X a fresh 'repo sync' now takes 9.0s.  Of this time 5.0s is spent
updating the database (mostly running lots of inserts -- about 130,000 of those).
The other 4.0s is mostly related to overhead in the XML parsing.

By dropping the feature indexes before the update and recreating them after
the time spent updating the database drops to 4.3s (14% reduction in time spent).

By not having these indexes at all we can update the database in 3.5s (30%
reduction in time spent).

[git-p4: depot-paths = "//depot/main/Camel/src/as-mod/ActivePerl-PPM/main/": change = 295449]